+++ src/main/java/global/util/Date.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 
-public class DateUtil {
+public class Date {
     public static LocalDateTime TODAY = LocalDateTime.of(LocalDate.of(2024, 12, 17),
             LocalTime.of(10, 30, 00));
     public static final LocalDate CHRISTMAS = LocalDate.of(2024, 12, 25);
